+++ index/src/score/mod.rs
@@ -6,7 +6,7 @@ use std::collections::HashMap;
 /// Stores the scores of each document.
 #[derive(Default, Debug)]
 pub struct DocumentsScores {
-    inner: HashMap<usize, f64>,
+    inner: HashMap<u32, f64>,
 }
 
 
@@ -21,7 +21,7 @@ impl DocumentsScores {
     /// If the document is present in the struct, it adds to its current score.
     /// If the document is not present in the struct, it adds the document with
     /// the specified ID and score.
-    pub fn add_score(&mut self, doc_id: usize, score: f64) {
+    pub fn add_score(&mut self, doc_id: u32, score: f64) {
         *self.inner.entry(doc_id).or_insert(0.0) += score;
     }
 
@@ -29,15 +29,15 @@ impl DocumentsScores {
     ///
     /// If the document is present in the struct it returns its score.
     /// If the document is not present, it returns 0.0.
-    pub fn get_score(&self, doc_id: usize) -> f64 {
+    pub fn get_score(&self, doc_id: u32) -> f64 {
         *self.inner.get(&doc_id).unwrap_or(&0.0)
     }
 
     /// Returns a vector of the documents with the top n scores. 
-    pub fn get_top_n(&self, top_n: usize) -> Vec<(&usize, &f64)> {
+    pub fn get_top_n(&self, top_n: u32) -> Vec<(&u32, &f64)> {
         let mut elements = self.inner.iter().collect::<Vec<_>>();
         elements.sort_by(|a, b| a.1.partial_cmp(b.1).unwrap().reverse());
-        elements.into_iter().take(top_n).collect()
+        elements.into_iter().take(top_n as usize).collect()
     }
 }
